# Make the tests run faster