--- a/meilisearch/tests/common/server.rs
+++ b/meilisearch/tests/common/server.rs
@@ -2,7 +2,7 @@
 
 use std::marker::PhantomData;
 use std::path::Path;
-use std::str::FromStr as _;
+use std::str::FromStr;
 use std::time::Duration;
 
 use actix_http::body::MessageBody;
@@ -345,6 +345,7 @@ pub fn default_settings(dir: impl AsRef<Path>) -> Opt {
             // memory has to be unlimited because several meilisearch are running in test context.
             max_indexing_memory: MaxMemory::unlimited(),
             skip_index_budget: true,
+            // We cannot use more than 1 indexing threads otherwise some tests become non-deterministic
             max_indexing_threads: MaxThreads::from_str("1").unwrap(),
         },
         experimental_enable_metrics: false,