… meilisearch/tests/auth/mod.rs
… meilisearch/tests/documents/add_documents.rs
… meilisearch/tests/dumps/snapshots/mod.rs/generate_and_import_dump_containing_vectors/1.snap
… meilisearch/tests/dumps/snapshots/mod.rs/generate_and_import_dump_containing_vectors/2.snap
… meilisearch/tests/logs/error.rs
… meilisearch/tests/search/errors.rs
… meilisearch/tests/search/mod.rs
… meilisearch/tests/search/multi.rs
… meilisearch/tests/search/pagination.rs
… meilisearch/tests/search/restrict_searchable.rs
… meilisearch/tests/search/snapshots/distinct.rs/distinct_at_search_time/succeed.snap
… meilisearch/tests/settings/errors.rs
… meilisearch/tests/swap_indexes/errors.rs
… meilisearch/tests/tasks/errors.rs
… meilisearch/tests/tasks/mod.rs
… meilisearch/tests/vector/mod.rs
… meilisearch/tests/vector/rest.rs
… meilisearch/tests/vector/settings.rs
… meilisearch/tests/vector/snapshots/mod.rs/add_remove_one_vector_4588/document-added.snap
… meilisearch/tests/vector/snapshots/mod.rs/add_remove_one_vector_4588/document-deleted.snap
… meilisearch/tests/vector/snapshots/mod.rs/add_remove_one_vector_4588/settings-processed.snap